requirements.txt
PyYAML==3.13
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-bigquery==1.24.0
ibm_db==2.0.9
retry==0.9.2
pandas==0.25.1
pyarrow==0.17.0
google-cloud-secret-manager==1.0.0
Flask==1.1.2
gunicorn==20.0.4